Something in X-motion is broken
There is a bug and I can't debug it due to the Full-screen nature of my
game. I'm going to move on for now while continuing to think of what it
could be.

Next step: Start making things that the character can jump on.
Step on Deck: Play with switching gravity left/right/up/down